Fix handeling ASCII and unknown bytes
Adding the value of a bytes bigger than 127 (ASCII) to the result
adds unknown bytes which could have been a part of a seqence or anything else.
We can't determine the type of non-ASCII bytes because the data needed to do that
is lost when the data was damaged.
The new implementation only adds the replace charecter and moves on to the next byte.