New implementation with performance improvement of 84%
Encoder::Converter is used right now to change the encoding after I found
out that the data can be repaired by just re-encoding the data to cp1252
and using it as utf-8! The new implementation yielded an improvement of more than 84%, sweet!

The reason this works is because when the data gets saved in a utf-8 database
with the wrong option for encoding (for example: cp1252, as it's the defualt in PHP),
the database tries to convert the utf-8 data from cp1252 to utf-8. This leads to the data
misencoded.

This can be further explained in the follwoing way (beginning and ending encoding is ALWAYS utf-8):

  misencoding    : treat data as cp1252         => convert from cp1252 to utf-8
  fixing-encoding: convert from utf-8 to cp1252 => treat data as utf-8

So basically fixing the encoding is just a matter of reversing the process of breaking it.